MetadataNotebook: 1) Added "New..." menu item
ClickableList:
1) Added cool icons
2) Fixed bug in which item is selected when
removing/adding items

MetadataPane:
1) Now supports semantic type defs in the file
2) Added cool icons
3) Finished the new handling of what happens when the
element found in the template is not found in the file.
Now it waits to be told to create a blank table before
it creates XML elements to correspond to the template.
4) Fixed bug in handling of changing the Tabs menu
when tables are blank and haven't been created
5) New method makeNode() uses reflect api to create
unknown named nodes with ease and in a general manner
6) Fixed bug in TablePanel's initialization of
isTopLevel. Now it handles null values appropriately.

ClickableTable:
1) Switched around menu items
2) Streamlined code to use MetadataPane.makeNode()
which saves some lines, adds clarity.
3) Resolved conflict with Curtis' changes

NotePanel:
1) Added cool icons

TemplateParser:
1) Added method getRoot() to get the template document
easier.

Icons:
1) Added cool icons to signify whether there is
metadata present or not.