NEW - GotoEditor: 1) a class that manages the editing behavior of the buttons in the "Goto" column
NEW - GotoRenderer:
1) a class that manages rendering buttons appropriately
in the "Goto" column of the tables

VariableComboEditor.java:
1) Massive overhaul that allows the tables to use
either a JTextField or a JComboBox depending on whether
the row is for a reference attribute
2) Egads, tables are not designed to do this at all
3) Added support for changing the xml nodes if the
editor is a JLabel (didn't work previously)

MetadataPane:
1) removed helper class TableButton, added it to
GotoEditor instead, as this makes more sense now
2) killed old handling of tableChanges that altered
the node tree changing, which is now in
VariableComboEditor.java
3) made MetadataPane independent of its frame
4) made the two tables in TablePanel into one --
this was disgustingly difficult.
5) added tooltip compatability to like everything

ClickableTable:
1) overrides of the getToolTipText method to allow
appropriate tool tips for the particular cell that
the mouse is over.

NotePanel:
1) fixed minor annoying bug that had to do with the
preferred size of notepanel being larger in width
than the rest of the tablepanel causing the tables
to resize when notepanels are displayed.

Template:
1) added more descriptions to handle the new tooltips

MetadataNotebook:
1) minor changes to handle new modularity of
MetadataPane, including constructor that doesn't need
a TemplateParser argument.