GotoRenderer.java: 1) changes for better performance based on suggestions from the jdk 1.4 DefaultTableCellRenderer API
ClickableList.java:
1) changes to certain uses of DOM Elements when I
could have been manipulating OMEXMLNodes instead

VariableTextAreaRenderer (NEW):
1) changed VariableComboRenderer to handle only cells
that should have comboboxes... this renderer signifies
with a snazzy icon that a cell will be edited by a
textarea instead of a textfield or combobox

VariableComboRenderer:
1) total overhaul, most of the mechanics of this are
now in MetadataPane. This keeps the renderer code
clean and simple and avoids nasty mishaps/bugs.

MetadataPane:
1) big cleanup of places that I was using Elements
instead of OMEXMLNodes

2) added code to override getRenderer or whatever it
is in ClickableTable so that depending on the row
the renderer for the cell will be appropriate

3) switched around where the internal defs are read
for better performance, since it shouldn't be re-doing
the internal defs every time it remakes the GUI.