VariableTextEditor (NEW): 1) an abstract super class for the TextComponent editors for table cells. Takes care of most of the listening.
VariableTextAreaEditor (NEW):
1) extends VariableTextEditor for TextAreas
2) changes the stopping of editing to happen only
if the textarea loses focus, instead of when the mouse
leaves the textarea's bounds. much better this way.

VariableTextFieldEditor (NEW):
1) extends VariableTextEditor for TextFields

VariableComboEditor:
1) got rid of all the textfield/textarea junk which is
now in seperate classes, as specified above. Now this
class only handles the comboboxes

ClickableTable.java:
1) override of getCellEditor method... by doing it this
way we segment the various types of editing components
we wish to use and the code is much more elegant/
readable

MetadataNotebook:
1) fixed bug with "New" menuitem that it failed to ask
if you wanted to make a new file without saving the
previous one when changes were made

MetadataPane:
1) changes resulting from the migration to seperate
CellEditor classes
2) .ome files are now searched for the first pixel image
just like .tiff are... this is kind of hacky code