Browser: 1) made MetadataNotebook the new way to edit/view metadata when the metadata button is pressed.
Notebook:
1) Normal view now starts on the first Image tab, not Project.
2) Dropdown combo boxes are now more descriptive for internal
references based on a new attribute "RefVars" in Template.xml
For now, these boxes do not reflect any user changes to the
attributes immediately, if for no other reason, that's a lot
of operations to do every time anyone changes anything. Perhaps
I'll add an "Update References" option eventually or something.
3) Buttons are now the default size since I don't set preferred
size anywhere anymore.
4) Added a file LOCIDefs.ome which holds "custom" elements like
LaserCoordinates and defines them for use in LOCI.
5) Fixed the problem with the "Goto" buttons when there was no
scrollbar. Should scroll better on ones with the scrollbar as
well since the Layout was reworked and the goto stuff was thus
outdated.
6) Made a new menu Options which allows the user to display the
ID attributes. Currently these attributes are editable, which
is somewhat buggy. Technically any time one of these are editted
the GUI should be remade.