Rewrote testing framework.  To use, run:
java loci.formats.test.ReaderTest /path/to/data/folder

Be sure to allocate plenty of memory (at least 512 MB), or you will get false
positives.  Note that data-files.txt is ignored, so only files in the specified
directory (or its subdirectories) will be tested.  A summary of the test results
is written to stdout in real-time; more verbose results are written to a logfile.