Avoid "prefix <...> not bound" error when transforming XML
When transforming XML via XSLT, all used namespaces have to be declared.
In other words, if you have a tag "<svi:Custom>", there needs to be
an attribute "xmlns:SVI" in the root element.

If such an attribute does not exist, the XML is malformed. To avoid an
exception, the tags are transformed to contain an underscore instead of
a colon. As these tags are typically in the CustomAttribute section (which
is filtered out anyway), no harm is done by this change.

This works malformed XML generated by Huygens.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>