OMETiffReader now populates its' MetadataStore in such as way as to produce valid OME-XML (hopefully).  BinData elements are not populated; instead, a MetadataOnly element is inserted under any Pixels that do not have at least one TiffData.  See #556.
The only problem with this is that OMETiffReader.getMetadataStore() no longer returns a MetadataStore that can be used with an IFormatWriter (assuming, of course, that an IMetadata was passed to setMetadataStore).  The current solution is a new OMETiffWriter API method: 'MetadataStore getMetadataStoreForConversion()'.
This will return a MetadataStore of the same type as 'getMetadataStore()', but that contains the required BinData elements.