Fix problem where custom LUTs override grayscale mode.
This happened because in some cases, the ImageProcessors already have
custom LUTs attached. We extract the LUTs and add them to a list, for
Colorizer to apply as appropriate, but in the case of grayscale,
Colorizer merely nulls the LUTs and does nothing, assuming that the
ImageProcessors are grayscale.

The fix is to revert the ImageProcessor LUTs back to the default,
prior to invoking Colorizer, so that grayscale mode works properly.