The 'uuids' field is no longer static.  See #639.
The only reason to have 'uuids' be static is so that multiple writer
instances could access the same cache of UUIDs, e.g. if you were to
write a single multi-file dataset using multiple writers.  However, we
have already documented that this is *not* the way to write multi-file
OME-TIFF datasets, since other metadata will be lost.