Pervasive and tile based option support. (See #4509, #4510)
Adjusted the JPEG 2000 codec option stack so that default options may be
set on the JPEG 2000 writer and that those options are used as the basis
for each compression action. Also added tile based properties (currently
unused) to the codec options that mirror the method arguments of JAI's
ImageWriterParam.setTiling() method.