Extended codec options for read operations. (See #4509)
In order to support sub-resolution retrieval additional (read-only)
options were added to the JPEG 2000 codec options to be used as the
basic for all decompression actions. Also added, in order to adequately
support the manipulation of the resolution (also known as decomposition
when writing) additional options were added with compression in mind.