Do not adjust the tile width if it is larger than the image width.
Some TIFF files legitimately use a tile width that is larger than the
image width and use padding bytes to make up the difference.  Thus the
tile width should not be adjusted if the image width is smaller than the
tile width, because the padding bytes will not be properly stripped out.