Fix JPEG-2000 pixel type detection.
Only one byte is used to encode the number of bits per component, not four.
The remaining three bytes encode the compression type, whether or not
the colorspace is known, and whether or not the file contains
intellectual property rights information (respectively).

Also, follow the JPEG-2000 specification in handling this byte, so that
files with 8 < bits per component < 16 are handled properly.

See also section I.5.3.1 of http://www.jpeg.org/public/15444-1annexi.pdf