Add new metadata keys that record the image orientation.
The keys are "Reverse X orientation" and "Reverse Y orientation".

These keys should be used if the orientation must be corrected, as we are
*not* doing this automatically in Bio-Formats.  Upon closer
investigation, I think it is better to record the metadata and preserve
the images as they are stored and displayed in LAS AF, rather than
automatically correcting the images.

See #5000.