Prevent an exception from being thrown if the 'gzip' flag is incorrectly set.
Some files' metadata will indicate that the pixel data is
GZIP-compressed, even if this is not the case.  To allow the pixels to
be read, we now unset the 'gzip' flag if the pixel data cannot be
GZIP-decompressed.
(cherry picked from commit fc42142a737fc51ac0f66c1c889b9604c3b8d2f7)