Make sure that we are opening a thumbnail of the correct size.
For efficiency, we try to open the thumbnail of the smallest series.
However, the smallest series may not have the correct aspect ratio, so
it is important to check that the dimensions of the thumbnail that we
are trying to open will match the expected thumbnail dimensions.
(cherry picked from commit e1388b284230a1c22d7b619b5de7136829441aef)