Updated BFITK MetaDataDictionary storage types
In BFITK-pipes the ITK MetaDataDicitonary is used
to store information about the image.  However, the method
for extracting data from the dictionary is typed - thus there was no known
way of finding a value, even given its key, without knowing the
value's type.

All values are now stored as strings and decoded when retrieved from
the dictionary.

The code on both Java and C++ sides of the pipe were adjusted to
remove tracking of type, except when extracting several specific
dictionary entries.  A generic extractor method was added for
convenience to the native code.