Add POM descriptors for external dependencies.
These are not used directly, but are copies of what lives in the Maven
Nexus repository. For example, in components/ome-io/pom.xml,
ome:omero-client:4.3.1 is declared as a dependency, so this JAR needs to
live in the repository, along with its own POM describing its
dependencies. In this way, dependencies can be discovered transitively.
(cherry picked from commit 54bea0a6aac9ff47d95ccad6ffa689815e82f0d4)