Reverted BFITKPIPE CMakeLists -platform hack
After testing on Linux it was discovered that the platform hacks
in fact did not work and did not make BFITK more portable.

This was decided to be a moot point anyway as, whether they worked or not,
the LD_LIBRARY_PATH (or equivalent) would need to be set for ITK to be found.

Thus the CMakeLists.txt file will be reverted to that immediately before the changes.