Reclassify Brian Loranger as a contributor.
This is because in Maven, "developer" is synonymous with "person you
would email about the project" so usually means only active developers.

I also removed the email addresses for Brian and Eric, since the
contributors are listed mostly for credit, not for people to contact
them (and the email addresses will probably expire anyway).