Moved CPPWrap from Bio-Formats to SciFIO
As SciFIO is lighter than Bio-Formats, it makes more
sense to wrap SciFIO with CPPWrap by default.  If users wish to wrap the
full bio-formats library, it can be done with manual CPPWrap use.

In addition to copying of files, the pom's for both project were
updated.

Also, the demo programs for CPPWrap - showinf and minimum_writer - had
their jar dependencies corrected to the SciFIO jar.