Undeprecate MetadataTools.createOMEXMLMetadata.
This method is incredibly concise and useful, and the deprecation
warnings were annoying. The only reason it was deprecated was due to
thread safety, since it kept a shared OMEXMLService instance. So for
this one method, I changed it to create the service during the method
call, to avoid any threading issues. It should be totally thread-safe
now, and more convenience than creating a new OMEXMLService.