Add another method signature that allows us to write pre-compressed pixels.
By setting nChannels to the number of RGB channels and setting
'copyDirectly' to true, you can write pre-compressed data to a TIFF
file.  If the 'copyDirectly' flag is set, then all of the pixel data
manipulation is bypassed and whatever was passed in is written directly
to the TIFF file.

See #6751.
(cherry picked from commit 48dc77dbcc44273f5ad9adf340ae707316691caf)