Fix bug in zip reader.
The issue was that RandomAccessInputStream(String) will convert a zip
file into a ZipHandle using Location.getHandle(String), but the
ZipReader was assuming that RAIS would be feeding it raw data from the
zip file itself.

The fix forces the RAIS to create a raw handle rather than a ZipHandle.