Allow the top level configuration file to be given explicitly.
This allows us to keep multiple configuration files in a directory, so
that we can easily test against different subsets of the data repository.

In particular, it makes it easy for the "Bio-Formats full repository data test"
job in Hudson to test different directories each day of the week without us
having to chmod 777 the data directory and copy the configuration files around.