Ensure that Image.AcquiredDate is preserved.
Calling MetadataTools.populatePixels(...) now automatically sets a
default Image.AcquiredDate.  For most formats, this does not cause a
problem, but for OME-TIFF it causes any existing values of Image.AcquiredDate
to be overwritten by the file's timestamp.

Thanks to this fix, the original values of Image.AcquiredDate are not
overwritten.