Add utility to generate the dataset structure table.
'ant gen-structure-table' in components/bio-formats will automatically
generate the table and place it in
components/bio-formats/doc/DatasetStructureTable.wiki.

The class that does the real work is
loci.formats.tools.MakeDatasetStructureTable.  Arguably, this should
have gone in the autogen component, but as the code is quite simple
and it is a useful example of how to use the new
getDatasetStructureDescription() method, it seemed better to put it in
bio-formats.

See #6867.