Ensure that a LUT is available once setId returns.
Previously, some readers for formats which stored one LUT per plane
would return null on calls to get*LookupTable() that occurred prior to a
call to openBytes(...).  Now, get*LookupTable() will return the LUT of
the 0th plane when called immediately after setId(...).

Most importantly, this means that we can eliminate the call to openBytes
in ChannelFiller.setId(...), the sole purpose of which was to ensure
that the LUT for plane #0 was available immediately.

Closes #7009.