Large-scale refactoring to allow tile-based conversion.
ImageConverter now converts planes one tile at a time, if the number of
pixels per plane exceeds 2**24 (4096 x 4096) and the output format is
TIFF or OME-TIFF.

In order to accomplish this, quite a bit of refactoring was necessary.
There are now separate methods to accomplish plane conversion, pixel
scaling, and color model conversion.

Closes #6971.