Remove Bio-Formats extension logic
It was originally added so that tools like the FARSIGHT Nucleus Editor
had a way to obtain the list of Bio-Formats-supported file format
extensions without needing to directly link to the Bio-Formats codebase.
This is necessary for tools that interface with Bio-Formats/SCIFIO via
via the BF-ITK plugin, since they cannot link to Bio-Formats directly.

Now that SCIFIO is a separate, BSD-licensed component, it is more
feasible for applications like FARSIGHT's Nucleus Editor to bundle the
SCIFIO C++ and Java components and link to it directly to query the list
of available extensions. This work is not done, but would be fairly
straightforward.