Reverse the usage of .ics and .ids files
This fixes the very, very stupid bug alluded to in the commit message
for 51e4fbdbf5ebe5d6d8a325d45741f0a0de7adb94.  Previously, the .ics file
contained the pixel data and the .ids file contained the metadata.  This
is the opposite of what it is supposed to be.