Fix channel count bug introduced in f8c4b76
The lookup table component count needs to be reset on every call to
setId.  Since we only really care about not calling
MetadataTools.populatePixels(...) on files that have already been
initialized, there is no reason not to let everything else run on every
call to setId.

This should allow the tests to pass again.