JPEG: Ignore suffixSufficient if 'open' is unset
This allows JPEG files to be opened even if 'setAllowOpenFiles(false)'
was called on the corresponding reader.  Note that non-JPEG files with
the '.jpg' extension will *not* open correctly in this case (though they
would not have been opened correctly anyway).

The reason for not simply changing 'suffixSufficient' is that the suffix
is not sufficient in general; Canon RAW and and DICOM in particular can
have the '.jpg' extension.  By overriding isThisType(String, boolean),
we keep the correct type detection for when 'setAllowOpenFiles(false)'
was not called.

Suggested by Lee Kamentsky and Curtis Rueden.