Add extra check for using the C index in filenames
Sometimes, the .htd file will specify (via the "Waves" flag) that
channel indices should be included in the TIFF filenames, even if only
one channel is present.  We now respect that, instead of assuming that
the index will not be present if SizeC == 1.

Closes #8013.

To test, verify against QA #4207/#4208 and run:

ant -Dtestng.directory=$DATA/cellworx test-automated