Add safeMultiply method for longs
This allows for checking 64-bit overflows in addition to 32-bit
overflows.

Note that the safeMultiply(long...) method algorithm is a bit different,
since we cannot promote the running total to the next highest primitive
type (as Java has no 128-bit primitive integer type). Instead, we count
bits of the two operands and throw the exception if the multiplication
would overflow. This method is fast, but not quite as fast as the
safeMultiply(int...) technique, so we leave the 32-bit algorithm alone.