Fill in missing units for the C axis
Some software (notably Huygens) does not feel the need to write a unit
for the C axis.  This causes our poor ICS reader a great deal of
confusion, as the index into the array of units is now wrong.  So now we
check for a missing unit, and add in "nm" for the C axis as appropriate.

Reported by Glen MacDonald.