Adjust the Y coordinate before flipping the plane
In general, it is not sufficient to just read the user-specified tile
and then flip it upside-down - this will only work if the tile height
is equal to the full image height.  In order to read tiles that are smaller
than the full image, we must reverse the Y coordinate before reading.

Fixes several errors noticed by BIOFORMATS-omero-openbytes.