Migrate ImgLib2 SCIFIO I/O classes from imglib2-io
This code previously resided in the ImgLib git repository:
  http://github.com/imagej/imglib

Specifically, it was the imglib2-io module (in imglib2/io subtree).
However, as part of the SciJava effort (http://scijava.github.com/),
we are working together to create a single common stack of imaging
software in Java. We decided that ImgLib2 would be at the bottom of
that stack, and that SCIFIO would depend on ImgLib2 to provide
higher-level I/O functionality. These classes are precisely that.

Hence, it makes sense to provide ImgOpener and ImgSaver as part of
SCIFIO core, rather than in a separate module, as long as SCIFIO is
already going to depend on ImgLib2.