Working implementation of ImgSaver
Reworked ImgSaver to clarify the execution paths to "writePlanes", the
functional workhorse of the class.

The point of Metadata population varies slightly based on whether or not
a Writer is provided to the saveImg methods (the public entry point).

With the corrections to Metadata population, the ImgSaver was verified
as capable of actually producing image output via Bio-Formats (although
it has not been subjected to any comprehensive testing at this point).