Added handling for special Dimension order
Added a check to ImgSaver for saving an image with multiple axes of
length > 1 that bracket a known axis, e.g.:
XYUZGTC

Where U and G are unknown. If unknown axes are adjacent, their lengths
are multiplied and forced to the "C" dimension. When unknown axes are
not adjacent, there is no way to compress the information to 5 planes.
Saving such an image is not currently supported by ImgSaver.