ImgSaver deletes old, existing files
Once ImgSaver is sure it will be able to send data to Bio-Formats for
saving, if the file being written to already exists it is deleted.

This is because of Bio-Formats' default appending to existing files. If
the API is changed to allow for overwrites, the handling in ImgSaver
could change as well.