Refactored AxisGuesser to only compress to C axis
The AxisGuesser guessImgLib method, which is used in ImgSaver to attempt
to compress the axes of an ImgPlus to a 5D structure compatible with
Bio-Formats, now will only compress unknown axes to Channel, and will
only substitute unknown axes for channel.

Thus:

XYUCUZT and XYUZT are valid but
XYCUZUT is not

Previously, the guessImgLib method would attempt to compress to an
arbitrary axis (and was non-functional in some cases).