Added isCompressible(Img) method
The isCompressible method can be used by higher level applications to
determine if the Img or ImgPlus object they would like to save will
be mangled during the saving process.

ImgLib supports axes that are not currently known to Bio-Formats, but
unsupported axes can potentially be compressed to the known Channel
axis. Note that this compression does change the structure of the
Img/ImgPlus.

isCompressible returns true iff the ImgPlus object will lose unknown
axes to compression.