Corrected isCompressible logic to check for U axes
isCompressible now does a first scan of the ImgPlus structure to
determine if there are unknown axes of size >2 (which would need to be
compressed).

If there aren't, it returns false immediately without needing to run
through the AxisGuessing algorithm.

If such an axis is found, then the AxisGuessing algorithm is run to
determine if the axes can be compressed, and a result is returned as
appropriate.

Note that there may be additional unknown axes of size >2 that can NOT
be compressed. In this case, isCompressible will return false, but the
ImgSaver will likely throw an exception later when attempting to save.