Clean up and unify ImgLib2 I/O sources
This reconciles changes made in imglib.git with those in scifio.git.

The only remaining differences between the sources are the file headers
and the package statements.

The ImgLib codebase uses tabs for whitespace, so for now, to be
consistent and enable easier diffing, the tabs are left in place here.