BMP: ignore the compression setting when possible
If the compression value is set to something other than 'uncompressed'
(0), then we first check to see if there are enough bytes in the file to
contain the uncompressed image.  If there are, then we assume that the
image data is uncompressed, as some programs think it proper to misuse
the compression value.