Updated ant to incorporate external findbugs
The ant findbugs targets always required an external findbugs.jar, so
there is no reason to maintain a local findbugs-ant.jar - as
findbugs-ant.jar is intended to be used with a paired version of the
findbugs.jar, and thus the two are packaged together.

Instead, the findbugs task is conditionally defined when the
findbugs.home property has been set.