QTWriter: fix the duration calculations
The time scale is now set such that the duration values will be
interpreted as being in milliseconds.  The duration itself is also fixed
so that it is less subject to rounding errors, and the time-to-sample
table is fixed so that the milliseconds per frame are recorded (instead
of frames per second).

Note that the time-to-sample may still be very, very slightly incorrect,
which means that there may be a very small difference between the stored
duration and the actual playing time.  For example, if we have a movie that
contains 1800 frames at 30 frames per second, the duration will be 60
seconds (60000 milliseconds).  The number of milliseconds per frame is then
33.33...; however, we are limited to storing an integer
number of milliseconds per frame, so this will be stored as 33.  Thus,
the actual run time of the movie will be 33 * 1800 = 59400 milliseconds
(59.4 seconds), or 0.6 seconds shy of the stored duration.