TIFF: add special case for reading big long arrays
If we encounter an IFD entry that contains a large array of longs, then
instead of repeatedly calling readLong() on the RandomAccessInputStream
we now read the entire array at once from the RandomAccessInputStream.

This improves performance for very large TIFF files where the array of
tile offsets contains 4+ million entries.