Initial copy of loci.common to Loci-Legacy
This is the first of 2 commits to preserve backwards compatibility with
the legacy loci.common package.

Loci-legacy is not in a functional state with this commit. It is
simply a copy, with no changes, of all loci.common classes from the
scifio component to the Loci-legacy component. This will ensure
that the delegator classes will have the same history as the original
classes.