Data tests: order first by test name, then by file
This means that we now run all of the tests for a particular file
together, instead of running e.g. all of the BufferedImage tests for
every file before all of the byte array tests for every file.  By so
doing, we drastically reduce the number of calls to setId, which means
that there is a substantial performance improvement.

See #8732.