Teach the Colorizer to respect color and wavelength metadata
Now when the "Colorize" option is selected, we will not just blindly
assume that the first channel is red, the second is green, and the third
is blue (in the absence of defined lookup tables).  Instead, we first
check the value of Channel.Color, and then the value of
Channel.EmissionWavelength and set the color accordingly.  This is very
much like what OMERO does.

Closes #9191.