ImportProcess: return the name of the current file
...not the name of the file that was initially passed to the reader.  In
the vast majority of cases, the two files are equal.  However, when
reading from OMERO, it is very important to use the name of the current
file as the name of the initial file will contain the user's login
credentials.

Closes #9440.
(cherry picked from commit 63b95adad4018f1bd04735d9ce3d961c7192f508)