PrairieReader: add support for inverted X/Y coords
This works by checking the CFG file for the
xYStageXPositionIncreasesLeftToRight and/or
xYStageYPositionIncreasesBottomToTop properties being set to True, and
then inverting all X and/or Y positions respectively if so.

With this change, Prairie data with X and/or Y coordinates inverted will
still stitch properly using Fiji's "Grid/Collection stitching" plugin.