Moved sixteenbitlossless to bottom of testng.xml
The potentially troublesome 16bitlossslessjpeg2000 test was moved to the
bottom of the testng.xml, to attempt to ensure it will be run last (as
it may cause memory fragmentation).

Note that this will be accomplished when running the tests directly via
testng. However, ant and maven seem to order their tests differently, on
the principle that unit tests should be independent (which technically
the tests themselves are. The ordering being exclusively due to memory
issues).