Created checkValidId method in Location
ome.scifio.io.Location now has a checkValidId which will check for an
existing IRandomAccess mapping to the provided id, or if no mapping
exists will attempt to create/close a new IRandomAccess using the id.

An IOException is thrown if unsuccessful.

A legacy delegator method to mirror checkValidId was also added to the
legacy loci.common.Location.