Removed ZipHandle close hack
ZipHandle, upon closing, would query whether a mapping existed for the
current ZipHandle instance. If so, the close method would do nothing (to
avoid destroying a mapped ZipHandle, which could still be returned -
although invalid - by future map requests).

This hack appeared to be used in conjunction with ImageReader logic for
attempting to open/close handles to check for valid IDs. It is no longer
necessary given centralized checkValidId logic in the Location class.