Removed Region casting
loci.common.Region is one of few legacy classes allowed to extended its
ome.scifio equivalent - in this case due to the presence of public
fields in the API.

Because of this direct extension, the intersection method was casting
the result of ome.scifio.common.Region#intersection which was
compile-safe but resulted in a ClassCastException at runtime.

This casting has been removed in favor of a direct translation method.