ImageJ 1.x plugin: warn the user about Standard ImageJ being deprecated
Standard ImageJ is an obsolete format, but still there for legacy macros
and plugins. Therefore, it should not be used anymore.

To inform the user, special-case the Standard ImageJ format to output a
custom warning with a red "Warning!" label when in interactive mode.

This developer considered implementing a more general solution to
deprecate choices, but reminded himself of the YAGNI principle:
https://en.wikipedia.org/wiki/You_ain%27t_gonna_need_it
So: until the time when a more general solution is needed, the current
one should do nicely.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>