Load native libraries from a .jar
Instead of assuming that the user has libturbojpeg.* and libjpeg.* on
their java.library.path, we now use native-lib-loader:

http://github.com/scijava/native-lib-loader

to load the native libraries from the META-INF/lib/ directory of
turbojpeg-1.2.1.jar.  Currently turbojpeg-1.2.1.jar only contains the
Linux libraries, but it will be updated in the near future to include
Windows and Mac OS X libraries as well.

native-lib-loader-2.0-SNAPSHOT.jar was built from this commit:

https://github.com/scijava/native-lib-loader/commit/b930cb4921d4d23b8417c5fcd36695b0ff0c48a8