Added DimensionSwapper tests
Adds the DimensionSwapperTest class which includes a set of tests to
verify dimension order and sizes are as intended when calling
swapDimensions (setting input order) or setOutputOrder.

This also serves as a regression test for a bug that was discovered when
setting the input order after setting output order (and the plane sizes
not being updated).