Added color model support to itkBFImageIO
The itkBioFormatsImageIO plug-in now preserves LUTs when reading and
writing.

During the ReadImageInformation step, in ITKBridgePipes, if a LUT is
found it is fully piped back to the ImageIO plug-in and stored in the
ITK MetaData dictionary.

This dictionary is now cached in by the itkBioFormatsImageIO after
ReadImageInformation is called (the default ITK imageIO behavior doesn't
seem to guarantee its preservation).

When writing, each LUT value is sent back to ITKBridgePipes and used to
re-build the ColorModel.