Remove commented-out multiplication
This multiplication was introduced in 0599079, and as far as I can tell
that was always a bug.  It works fine if at least one of the following
is true:

  * only one tile
  * only one channel
  * non-planar channels

However, if we have a multi-channel, multi-tile image with planar
channels then SizeC times too many tiles are read into the image buffer.

See also github.com/openmicroscopy/bioformats/pull/211.