CoreMetadata: Simplify copying and cloning
Add copy constructors and clone() methods to replace the
existing copy method.  This permits core metadata of
unknown derived type to be safely cloned.  Copying is
performed directly using copy constructors.  This also
permits the removal of the generic copyCoreMetadata
method in ReaderWrapper.  getCoreMetadata in ReaderWrapper
may now simply clone() a copy of the core metadata, and
setId in DimensionSwapper can use the DimensionSwapper
copy constructor to do a copy to derive a copy of the
core metadata.