Reset ifdCounts on close; stopped tweaking IFDs
When a writer is 'closed', the current ifdCounts map has all present
entries reset to 0. This map is referenced in populateImage, but if you
change output files (multi-file dataset), the stored IFD count is used to
populate the TiffData elements in each series. Each subsequent file would
cause the old TiffData IFDs to be replaced with a new sequence beginning
at the old maximum.

Also, the IFDs for series 0 were being reduced by 1. This caused each IFD
in series 0 to be identical.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>