Fix implementation of CoreMetadata.clone()
This means that clone() should work correctly if CoreMetadata is ever
subclassed.

I assume that this is safe thing to do based upon this section of the
Javadoc
(http://docs.oracle.com/javase/1.5.0/docs/api/java/lang/Object.html#clone()):

"The method clone for class Object performs a specific cloning
operation. First, if the class of this object does not implement the
interface Cloneable, then a CloneNotSupportedException is thrown. Note
that all arrays are considered to implement the interface Cloneable.
Otherwise, this method creates a new instance of the class of this
object and initializes all its fields with exactly the contents of the
corresponding fields of this object, as if by assignment; the contents
of the fields are not themselves cloned."

...which to me reads as though 'super.clone()' will do exactly what 'new
CoreMetadata(this)' is doing.