PrairieReader: remove obsolete channels logic
We no longer care about which channel min/maxes exist per frame, as we
rely totally on the CFG metadata to determine the desired channels.

To be absolutely sure that all of our sample Prairie datasets have
channel metadata given in CFG, I did a quick check:

    find . -name '*.cfg' -print0 | xargs -0 grep -L channel_

And indeed they do. If we come across a dataset in the future with this
information missing, it would be pretty easy to synthesize based on the
per-frame channel min/maxes again, but until then, we don't need it.