xsd-fu: Set current language and add class properties property
The current language global permits the model classes to know which
language they should be using without having access to the
command-line options object.

The class properties (instanceVariables) details the language-
specific instance type, name, default value and comment for
each class member.  This removes the need for the code to
determine these values to be duplicated in the template for
each language and for each time they are needed which greatly
simplifies the template.