xsd-fu: Move Pojo template python functions into fu.py
The parent class (superclass type) is defined directly as a model
object property.  The parents dictionary is no longer passed to
genshi via xsd-fu for Pojo.template since it's no longer required
with the per-object and per-property properties.

The main rationale for this change is to simplify the template
and to also allow C++ code to chain up to its superclass (which
unlike java requires explicit knowledge of the type).