Split out main doc sections with bullets
We were already listing them as bullets, but then we also redundantly
linked to them in the introductory paragraph above as well.

Bulleted lists are often much easier to read and digest
(http://www.contentious.com/2004/09/17/bulleted-lists/), so let's just
stick to a single list with the description inline per bullet point.