A reader for Yokagawa CellVoyager datasets - WIP
Yokagawa CellVoyager microscope can acquire large areas composed of
several tiles. Each plane (1C, 1Z, 1T) is stores as one plain tiff file
per tile.

Apparently:

- The stitching is done border to border. There are no
subtleties in arranging the tile in space. Mistakes can be seen even in
the proprietary software. Maybe some registration is done prior to
saving; we could not assess for sure.
- The CellVoyager is a spinning-disk. Individual tiles are corrected for
uneven illumination, apparently using a circular gaussian, prior to
saving. This can be seen on densely labeled tiles by changing the display
min&max.

The metadata and tile positioning in space are stored in a first xml
file. The list of tiff files composing the dataset is stored in another xml
file. This reader expects to be pointed to a file in the directory containing
the xml files.

What it does now:
- Open 5D datas, stitched on the fly. That is: if a plane is made of 4
  fields of 920x920 pixels arranged in a square, it will return a plane
  of 1840x1840 pixels.

Current limitations:
- OME metadata is not set yet.
- Only reads the first Area of the first Well yet.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>