Update NetCDFServiceImpl and HDF-based readers for netCDF 4.3.19
There are a number of differences in the group/attribute names and
values returned by 4.3.19 when compared with 4.0.x.  Some of the names
contain spaces instead of underscores and/or extra characters; for image
data arrays, the endianness may be reversed (which was actually a bug in
4.0.x).

This completes ticket #6853.