xsd-fu: Rename templates to match commands
This is to make the templates match their actual purpose, since
the status quo is quite confusing with the names being quite
unrelated to the command being run.

- Pojo.template renamed to OMEXMLModelObject.template;
  template fragments renamed accordingly
- Enum.template renamed to OMEXMLModelEnum.template
- EnumHandler.template renamed to OMEXMLModelEnumHandler.template
- Enums.template renamed to OMEXMLModelAllEnums.template

As for the command name renaming, this makes the system much
more obvious and self-documenting, and also matches the
existing OMERO templates making it clear these are separate
from the templates used for the OME-XML model.