Add Log4jTools for MATLAB support
In order to match the behavior in MATLAB with 4.4,
`DebugTools.enableLogging("INFO")` etc. should continue
to work. Since log4j is shipped with MATLAB 2013, the
only way for this to function is for us to fallback to
configuring Log4j if Logback is not present.

Note: the use of `LOGGER.debug` in `ReflectedUniverse`
makes it unusable for the initial logging configuration.
Likely all uses of LOGGER should be removed from RU which
shouldn't be an issue since for each log statement a
`ReflectException` with the same message is thrown.