Move MetadataStore, MetadataRetrieve, and related classes to ome.xml.meta
This adds a dependency on loci-common to ome-xml; I would not expect
this to cause a problem.

Wrapper classes with the old package names are kept and are not
deprecated at the moment.  With the exception of MetadataRoot and
OMEXMLMetadataRoot (which were not present in 4.4.x), the old package names
are still used across the board.  For IFormatReader etc., we're going to have
to choose between using only the old package names and using only the new
package names - things like 'MetadataStore getMetadataStore()' mean that
deprecating current signatures and adding new signatures for the new classes
is not really possible.