ImagePlusReader: get OME-XML from ImportProcess
The ImportProcess takes care of caching the OME-XML string so that we
don't have to generate a new copy of the OME-XML every time we need one.

This change reduces memory use when BF.openImagePlus needs to create
many ImagePlus objects. Previously, every ImagePlus object would have a
separate copy of the OME-XML embedded in its FileInfo#description field.
With this change, each description field references the same string.