POM: depend on the correct ImageJ 1.x GAV
The nih.gov.imagej GAV is obsolete and deprecated. Furthermore, it uses
an incompatible artifact name, resulting in both "ij-" and "imagej-"
artifacts on the classpath of anything that seeks to use Bio-Formats as
a dependency.

Since the net.imagej:ij:* artifacts are now available in Maven Central,
there is no more issue with this dependency resulting in any reference
to the maven.imagej.net Maven repository.