cpp: xerces: Clean up construction, assignment and implicit conversions
The only implicit conversion is now to bool, with explicit get()
methods providing access to the wrapped xerces-c node pointers.
This removes abiguity when used in if statements, where it will
now properly evaluate as bool.  It was previously evaluating as
the wrapped pointer and comparing to null (which was effectively
the same).

Also fix construction and assignment so that Element will
correctly assign the correct parent Node pointers as well.