CZI: use a buffer size of 1024 in most places
We have to do a lot of seeking and opening new streams to read a small
number of bytes; the only place where large chunks of bytes are read is
when the pixel data is processed.  This should speed up setId times, as
we're not attempting to buffer 1MB every time a new stream is opened
(potentially millions of times for very large datasets).