ImportProcess: remove UI-level exception reporting
Since this is a private method that throws FormatException and
IOException, the caller is assumed to be responsible for any UI-level
logging of those exceptions.  In this case, Importer is the only class
that makes use of the method in question, and it already displays an
appropriate error message when an exception is thrown.

Fixes #11866.