xsd-fu: Use package namespaces for all code generation
Don't use a mixture of dynamic and static package names, which in
practice made changing the names impossible.  Use only the dynamic
names.

This change permits the model to be loaded once and reused
for generating multiple targets in a single xsd-fu invocation.
Add new omexml_model_all and omexml_metadata_all targets to
generate all model and metatada code instead of using the
separate targets.