Adding and OME units wrapper (using uomo as base)
This wrappers is as light as possible and not functional at present.
It does allow `ant clean compile-ome-xml` to build successfully.
While it uses uomo it could be changed another by hopefully just
changing the ome.units.* files and language.py in xsdfu.

The OMEXMLModelEnumHandler.template will need a little bit more work.