PGMReader: use whitespace instead of lines to delimit entries on file header.
The netpbm file formats has 2 or 3 entries on the header delimited by
whitespace. The current implementation only recognized individual lines.
It is possible that there's no newlines in the entire file which caused
the reader to blow. In adition, this formats also allows for comments
anywhere before the raster data while currently they were only identified
if at the start of a line. Fixed all this by making use of StreamTokenizer.