PrairieReader: also parse ENV metadata files
Start with version 5.2, the CFG file was replaced with an ENV
that is similar conceptually, but with a revised XML schema.

It does contain many of the same key/value pairs, though, including
the ones in which we are most interested (channels, bitDepth, laserPower
and xYStageYPositionIncreasesBottomToTop in particular), so for now we
parse the metadata into the same config table used by the CFG file.