Use in-place caching by default in Memoizer
This commit sets the in-place caching attribute to true by default. The
immediate consequence is that all constructors now become usable (with only
the constructors using a directory input overriding this behavior). The
constructor using a boolean value now becomes obsolete and is removed. Unit
tests are added for each supported constructor using getMemoFile().