OMETiffReader: support BinaryOnly OME-TIFF refs
This commit changes the behavior of the OME-TIFF reader to support
extraction of OME-XML metadata from an OME-TIFF reference via the
BinaryOnly MetadataFile attribute.

According to a38bd0770eff839f2daf0289b10e5c6ae76afd5c, the BinaryOnly
MetadataFile is only supposed to reference an OME-XML document, not an
OME-TIFF file containing an embedded OME-XML document.

But according to the OME-TIFF specification [1], it is valid to
reference such an OME-TIFF.

Since supporting it is easy and very useful and already documented,
let's go ahead and do it!

[1] https://www.openmicroscopy.org/site/support/ome-model/ome-tiff/specification.html#storing-partial-metadata-blocks