Revert to using java.text.SimpleDateFormat for retrieving timestamps
Millisecond parsing in Joda has different results than with
Date/SimpleDateFormat.  If the number of ms digits in the format string
exceeds the number of ms digits in the date string, then zero padding
happens on the wrong side (e.g. 93 becomes 930 instead of 093).  If
the number of ms digits in the date string exceeds the number of ms
digits in the format string, then parsing fails.  See
https://github.com/JodaOrg/joda-time/issues/62.

Fixes #12117.