cpp: ome-xml: Add static create methods to model objects
Due to the fact that shared_from_this() isn't usable inside
constructors, it's not possible to call the update() method
in a constructor since taking references etc. of the object
being constructed will cause problems.  To avoid this, add
create methods which construct the object, call update(), and
then return a shared_ptr to the object.