BytesWrapper: new ReaderWrapper abstract subclass
In preparing Colorizer, another new ReaderWrapper subclass,
calls to "reader.getPlane" were not being intercepted since
there was only an implementation for the one abstract method
in FormatReader. This is a problem that other ReaderWrappers
are prone to. BytesWrapper is intended to replace the subclass
of any ReaderWrapper implementation which modifies openBytes
and friends with the intent that only one method need be
implemented.