Colorizer: New ReaderWrapper to apply a LUT
Colorizer makes use of the new BytesWrapper
subclass and the loci.common.ij package to
apply lookup tables to the byte array output
of openBytes operations. This is related to
the behavior of ChannelFiller and the similarly
named loci.plugins.ij.Colorizer class, but
allows more exact overriding of the behavior.
(For example, ChannelFiller manually checks
that the delegate reader is indexed and has
a LUT whereas Colorizer injects the fact that
indeed colorizing should take place.)

In addition, ImageConverter (i.e. ./bfconvert)
now takes an argument "-colorize $LUT" which will
be applied on writing.