cpp: bioformats: Use Boost.Filesystem for all system paths
The reason for this is to have strongly-typed paths.  The
Boost.Filesystem path class represents paths portably, so
will be useful for Windows portability.  It also ensures
that a consistent representation is used for paths,
filenames and extensions.  Additionally, it prevents all
passed strings requiring conversion to path objects to
do anything with them (which is quite common), since they
will now already be of the required type.