xsd-fu: Handle C++ lists and elements with const and non-const variants
Add non-const list getters.  These are an encapsulation violation but
are needed for strict compatibility with the Java implementation and
to get access to non-const element references.

Add element variants of argument and return types in property.py.  This
allows the templates to distiguish between element and container types
and support getting and setting of containers and container elements.