TIFF writer: automatically switch to BigTIFF when possible
If a BigTIFF extension is used, or we know that more than 4GB of data
will be written, then BigTIFF is used automatically.  This does not
attempt to switch to BigTIFF when compression is used and a BigTIFF
extension is not present, as we can't easily predict the actual size
of the output file in that case.

Fixes #6589.