ImageJ plugin - Exception when both 'group files' and 'open individually' options selected
A test case and steps to reproduce are linked from the ticket.
https://trac.openmicroscopy.org/ome/ticket/12941

Issues included:
- IndexOutOfBoundsException when both options selected in Main Dialog
- For certain file formats the regular expression and 'file contains'
file patterns were not being implied
- Entering incorrect values into the Dimensions setting of the
FilePatternDialog resulted in a 'File is not of a supported format'
error

Solution:
- IndexOutOfBoundsException resolved in ImportProcess by reading the
group files option. When both options are selected then group files is
not automatically the inverse of ungroup files
- For file formats that have a fileGroupOption of FormatTools.MUST_GROUP
the second FilePatternDialog is skipped and replaced with an ImageJ
message box informing the user why file pattern options are not
available, stating "Image specifications state that files of this given
format cannot be handled separately and must be grouped according to
supporting metadata". A new importer option flag was added to test and
store the must group condition