source-archive: Make tar source release
This script creates both a zip and tar source release.  The zip
file is unfortunately incompatible with some extractors, making it
break the superbuild on a number of platforms using libarchive to
extract.  This includes Ubuntu and BSD systems using the current
release of libarchive when using "cmake -E tar xfv" to extract the
zips (done in the superbuild).  Since fixing these platforms won't
happen in the short-term, making an archive they can extract is a
simpler solution.  The content is exactly the same as the zip.