octave and matlab: drop requirement on jar filename (issue #2018).
bfCheckJavaPath() was checking for the presence of a bioformats_package
or loci_tools jar files.  But filenames mean nothing and the required
classes for this to work may be in a file with any name.  So check that
we are able to construct a reader instead and if so, assume that it's
working.