octave and matlab: change logic to automatically find bioformats (issue #2018)
Automatically finding a jar file with the bioformats classes is limited to
filenames 'bioformats_package.jar' and 'loci_tools.jar' on the same directory
as the toolbox.  However, it is common to have such jar files with version
names.  Also, at least in Linux distributions there are conventions to where
jar files are available.  Add support for this two things.

Also, if we already have a status return value, there's no need to also
throw an exception sometimes.