Remove non-OME DLLs from the repository
As discussed with RichardMyers et al., having these opaque
binaries in a GPL'd codebase puts the OME Team in a precarious
position. It was decided to leave the status quo for 5.1.8
but to remove them early enough so that alternative delivery
methods could be found for the SlideBook6Reader users.

See:

 - https://github.com/openmicroscopy/bioformats/pull/1993
 - https://github.com/openmicroscopy/bioformats/pull/2217