Fix Bio-Formats importer bug for file grouped with the "Dimensions" option
The issue can be reproduced by opening a set of images grouped using a pattern
where the index of the first image is not zero. Instead adding the index of the
first image to calculate the index of the last image should be sufficient to
group all files together.

See http://forum.imagej.net/t/little-bug-in-bio-formats-importer/1461