cpp: ome-xml: Throw less exceptions in OMEModelObject::getChildrenByTagName
Rather than waiting for an exception to be thrown by the
Element RAII wrapper, check whether it will throw
beforehand.  This isn't strictly necessary, but it makes
the processing slightly more efficient and, more
importantly, makes it easier to debug the model code
since only important exceptions will be thrown.