Fixes finding BF_DIR if using symlinked executable
BF_DIR was resolved by a call to `dirname`, but in case the executable
is symlinked, for example, to /usr/local/bin, this fails. Unfortunately
there is no OS-independent way of resolving a symlink, so we try to use
`readlink -f` which should work for most of Unix except OS X. If that
doesn't work we use Perl (works on OS X), and if that doesn't work we
use the old behavior.

Resolves: openmicroscopy/bioformats#2540