OMEXMLReader: Handle BinData in non-Pixels elements such as Mask
The reader made the assumption that BinData was only found inside
Pixels.  However, it's also contained within Mask, and the reader
was stripping out all BinData content when initially reading the
OME-XML (it pre-processes it with a custom SAX parser before
creating the MetadataStore).  This change makes the stripping
selective, so it only occurs for BinData elements contained
within Pixels.