CV7000: do not overwrite the first action's pixels files
If multiple "actions" are defined, only use the files for the first one
to read pixel data.  All files for subsequent actions are ignored for
the purposes of reading pixels, but are logged and added to the used
files list.

This may be a temporary solution if we decide to read multiple actions
as separate fields.