File patterns: expanded channel axis detection
This expands the case for detecting a channel axis based upon "red",
"green", and "blue" in the file names.  Matching at least two of "r",
"g", and "b" is now sufficient instead of all three.

A pattern such as a_<red,green,greenyellow>.tif should now be detected as 3
channels.