Review Bio-Formats versioning policy for Bio-Formats 5.4.x and onwards
This commit makes official rules which have been almost in place for the
last patch/minor releases and should improve transparency for consumers.

- Precise the OMERO/Bio-Formats versioning decoupling point
- Use semantic versioning for the Bio-Formats API policy. This policy has
  been effectively in place since Bio-Formats 5.4.0.
- Increase the stringency of the serialization policy
- Add paragraph about the versioning of individual format changes

In terms of release, as per the usage of semver for the API, one expectation is
that it should be possible to have more minor releases of Bio-Formats
containing API backwards-compatible additions, format reader additions and/or
backwards-incompatible or serialization breaking format reader changes.