Fix Ant build.version property for non-SNAPSHOT versions
This commit fixes the behavior of the Ant calculation of the build.version
property to use `git rev-parse --verify HEAD`  independently of whether the
a SNAPSHOT or a release version is built. This implementation is in agreement
with the behavior of the Maven build number plugin and should fix the current
discrepancy for the value of Implementation-Build stored in the JAR manifest.