Update global metadata population so that it is displayed before series metadata
A null byte is now prepended to each global metadata key.
For applications such as ImageJ which merge all metadata tables
together, global metadata will then always appear before any series metadata,
as the series metadata table is deemed to be less useful (but worth
including for completeness).  The change was made in the reader instead
of ImageJ as this is likely not appropriate for other formats.