OMEXMLServiceImpl: remove Modulo annotations for validation
Fixes point 4 of https://trac.openmicroscopy.org/ome/ticket/12935

Setting "Label" on a Modulo annotation causes validation errors, because
it conflicts with the "Label" shape.  This only affects OME-XML
validation outside of the reader or after setId, as the optional
validation step in setId occurs before the Modulo annotations are added
to the MetadataStore.