ImagePro SEQ: support .ips metadata files
Fixes https://trac.openmicroscopy.org/ome/ticket/4180

If an .ips file is passed to setId, then it will be parsed and used to
group multiple SEQ/TIFF files together.  If the .ips file is present but not
passed to setId, then it will not be detected, as there is no back
reference from the SEQ/TIFF to the .ips file.