Disable snapshots for the unidata-releases Nexus repository
Following upstream deployment changes on the http://artifacts.unidata.ucar.edu
repository, the Ant build of Bio-Formats is seemingly broken during the
mvn-ant-tasks:dependencies step of the formats-gpl component. The Maven build
seems to be unaffected by these changes.

This commit tries to fix these issues by disabling the snapshots retrieval
from the Unidata repository. In practice, there should be no harm in doing so
as the only dependency we are consuming is the netcdf and we should not use
SNAPSHOTS with the unidata-releases repository.