TiffWriter: Don't split planes when using nonstandard SamplesPerPixel
The existing behaviour seems rather ad-hoc and potentially
dangerous.

- We ignore the SPP the user specified and write out something
  completely different, silently
- ChannelSeparator and bfconvert already provide the means to
  do this explicitly and intentionally
- While the aim of this appears to be as an aid toward greater
  interoperability with other software, it prevents the user
  writing out e.g. images with 2 or 4 samples per pixel.  If
  the user is writing out an OME-TIFF, the TiffData elements
  won't be able to represent the altered structure (the
  separated planes), and the file will not be readable as an
  OME-TIFF because the OME-XML metadata will not match the
  TIFF SamplesPerPixel

This logic also breaks a loop in the writer:
  saveBytes → prepareToWriteImage → saveBytes
Combined with tiling, this could result in calling saveBytes with
the wrong series index.  This removal makes the code easier to
understand, and removes some dangerous assumptions to make it
safer.

The default PhotometricInterpretation is adjusted to be correct
with the special case removed.