MicroManager: use PositionName key/value for storing the ImageName
Currently the Micro-Manager reader uses one of the TIFF filename for populating
the ImageName metadata. For typical replica experiments across multiple
positions (Pos0, Pos1,...), this results in duplicate image names.
This commit now parses the PositionName key/value in the metadata JSON and use
it if available to fill the ImageName.