Add cachgen command line tool for generating memo files
cachegen can process a single file, a directory, or a text file containing a
list of files.  Directory processing is less sophisticated that the
import candidates style used by the data tests, but this could be
considered for future work.  At the moment, Memoizer is the only reader
in the stack, but FileStitcher could also be added as an option.