Add IPyramidStore interface and OMEPyramidStore implementation
IPyramidStore specifies methods for getting and setting pyramid resolution
dimensions.  OMEPyramidStore implements IPyramidStore, and should be a
drop-in replacement for OMEXMLMetadata.

This might not be what we want in the end, but is a minimally invasive
way to provide pyramid metadata so that we can focus on the pyramid writer
itself.