Switch verbose TiffData parsing logging from DEBUG to TRACE
'showinf -debug' is required to show certain exceptions, but this can
result in hundreds of MB of output for OME-TIFF datasets with thousands
of planes.  This defers to TRACE all of the logging on which planes are
filled or cleared after each TiffData is parsed, which drastically reduces
the log size and initialization time.