Average a subset of recorded overlaps instead of using overlap list
A lot of trial and error indicates that the recorded overlap list for
each AOI does not indicate the total number of overlaps to be removed.
Individual overlap values are also sometimes incorrect by a significant
amount.

This takes the approach of averaging high-confidence overlaps to produce
a single value for each of X and Y, which are then applied to all tiles
in the AOI.