Update SubResolutionFormatReader to extend FormatReader
FormatReader now implements and uses currentCore() whenever the
CoreMetadata for the current series/resolution is used.  This allows
SubResolutionFormatReader to extend FormatReader, which means that
hundreds of lines of duplicate code are removed.

SubResolutionFormatReader continues to override all of the methods
specific to series and resolution handling, and uses a CoreMetadataList
instead of List<CoreMetadata> to store metadata internally.