Improve global metadata support
With the exception of AP_TIME and AP_DATE, all tags starting with
AP_, DP_ and SV_ have their next lines formatted as key = value. All these
pairs are parsed and added to the global metadata table.
Additional handling is retained for keys which can be mapped to the
OME metadata model.