TIFF: don't assume all IFDs have same XY size during non-sequential write
If setWritingSequentially(true) was not called, then writing each
resolution of a pyramid in order caused the previous resolution's
dimensions to be used.  This updates the IFD reuse logic so that only
tags which are not present in the 'ifd' argument are reused.  Use cases
where setWritingSequentially(true) is called will not be affected.