NDPIReader: populate both nominal and calibrated magnification
Support for the magnification in the Hamamatsu NDPI reader was introduced in
Bio-Formats 5.1.0. The original implementation makes use of the
Objective.Lens.Magnificant key in the metadata TIFF tag (65449).

The SourceLens TIFF tag (65449) also contains some integer values when it is
not equal to -1 and -2 to indicate ancillary images.

This PR proposes to adjust the way the OME metadata is populated to use:
- the value of the SourceLens TIFF tag as the nominal magnification
- the value of the Objective.Lens.Magnificant key as the calibrated magnification