Fix null pointer exception on close()
It is possible that classes will be rehydrated from memo files where
`allFiles` was not present and thus will be initialized as `null`.  If
this is the case, `close()` will throw null pointer exceptions when
called.  This may happen even for cases where the VectraReader is never
used.

Mitigation for scenarios where this bug is present are to remove all
memo files on the system and regenerate them.