CellWorX: add support for single site HCS variant
The number of well samples/fields of view per well is computed from the total
number of acquisition sites and the number of selected sites, determined by
XSites, YSites and the SiteSelection<index> entries in the HTD file.

For single site acquisitions (XSites=1, YSites=1), the SiteSelection1 key
might be absent from the HTD files. This commit updates the reader to assume
the single site is selected.