Deltavision: sanity check the panel count and fall back to old style header
Some files have the file type value set to indicate a "new" header format,
but actually have an "old" header format.  This causes the panel/series count
to be wildly inaccurate, which in extreme cases can trigger an OutOfMemory.
A basic check on the number of detected panels allows the reader to fall
back to parsing an "old" header instead.