Update DICOM dictionary and original metadata logic
There are a few compromises here. One is that the DicomAttribute enum
does not contain the entire dictionary, but a subset that is most likely to
be explicitly referenced in the reader/writer. The very large size of the
dictionary combined with Java's limitations on bytecode size means that
a single enum with the whole dictionary is not possible.

The other compromise is that the original metadata table does not contain
every tag in every file of a dataset, as that can quickly lead to memory
exhaustion (especially once the table is included in OME-XML).
DicomReader now has a "List<DicomTag> getTags()" method that can be used
to access the tag hierarchy directly, should an application need to do so.