Fix one more pixel offset calculation issue
Computing pixel offsets for JPEG and JPEG-2000 data is more involved than what the schema indicates.
In theory, we should be able to just read the length in each encapsulating item to determine the
length of the compressed stream. However, we have several non-WSI datasets that contain incorrect
item lengths, so this approach can't be trusted to work in general. Additionally, we have a few
WSI datasets that include 0xff4fff51 in the middle of a JPEG-2000 code stream. This is valid, but
confuses the scan for 0xff4fff51 denoting the beginning of the stream. The current imperfect
(but working) approach is to scan for the 4 magic bytes, and accept them as the start
of the stream only if a closing marker (0xffd9) was encountered between the previous offset
and the found magic bytes.