Fail fast if a file declared in the OME metadata is not found
The previous version of the code suggested a search feature which was never
implemented and resulted in a FormatException anyways with the caveat the
context of the original filename was partly lost.

This change should preserve the same behavior as previously but fail as soon
as a file in a multi-file fileset is not found with a more informative error
message including the UUID and the expected filename from the metadata.